+++ src/main/java/com/example/demo/review_like/application/dto/LikeCountDto.java
@@ -1,0 +1,4 @@
+package com.example.demo.review_like.application.dto;
+
+public record LikeCountDto(Long likeCount) {
+}
