+++ src/main/java/com/example/demo/review_like/application/dto/LikeStatusDto.java
@@ -1,0 +1,4 @@
+package com.example.demo.review_like.application.dto;
+
+public record LikeStatusDto(boolean isLiked) {
+}
