--- src/main/java/com/example/demo/review_like/dto/LikeCountDto.java
@@ -1,4 +1,0 @@
-package com.example.demo.review_like.dto;
-
-public record LikeCountDto(Long likeCount) {
-}
