--- src/main/java/com/example/demo/review_like/dto/LikeStatusDto.java
@@ -1,4 +1,0 @@
-package com.example.demo.review_like.dto;
-
-public record LikeStatusDto(boolean isLiked) {
-}
